+++ api/v1alpha1/cassandra_webhook.go
@@ -62,5 +62,9 @@ func (in *Cassandra) ValidateDelete() error {
 		return errors.New("cannot delete Cassandra service, termination protection is on")
 	}
 
+	if in.Spec.ProjectVPCID != "" && in.Spec.ProjectVPCRef != nil {
+		return errors.New("cannot use both projectVpcId and projectVPCRef")
+	}
+
 	return nil
 }