+++ api/v1alpha1/kafkaconnect_webhook.go
@@ -59,5 +59,9 @@ func (r *KafkaConnect) ValidateDelete() error {
 		return errors.New("cannot delete KafkaConnect service, termination protection is on")
 	}
 
+	if r.Spec.ProjectVPCID != "" && r.Spec.ProjectVPCRef != nil {
+		return errors.New("cannot use both projectVpcId and projectVPCRef")
+	}
+
 	return nil
 }