+++ api/v1alpha1/postgresql_webhook.go
@@ -63,5 +63,9 @@ func (r *PostgreSQL) ValidateDelete() error {
 		return errors.New("cannot delete PostgreSQL service, termination protection is on")
 	}
 
+	if r.Spec.ProjectVPCID != "" && r.Spec.ProjectVPCRef != nil {
+		return errors.New("cannot use both projectVpcId and projectVPCRef")
+	}
+
 	return nil
 }